src/test/java/com/springfwcourse/controllers/SetterInjectedControllerTest.java
package com.springfwcourse.controllers;

import com.springfwcourse.didemo.springframework.controllers.SetterInjectedController;
import com.springfwcourse.services.GreetingServiceImpl;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class SetterInjectedControllerTest {
    private SetterInjectedController setterInjectedController;

    @Before
    public void setUp() throws Exception{
        this.setterInjectedController = new SetterInjectedController();
        this.setterInjectedController.setGreetingService(new GreetingServiceImpl());
    }

    @Test
    public void testGreeting() throws Exception{
        assertEquals(GreetingServiceImpl.HELLO_WORLD,setterInjectedController.sayHello());
    }
}
